tests/wpt/web-platform-tests/orientation-event/deviceorientationabsoluteevent.https.html

<!DOCTYPE html>
<html>
  <head>
    <title>DeviceOrientationEvent attributes should be fired properly</title>
    <meta charset='utf-8'>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <p>If an implementation can never provide absolute orientation information, the event should be fired with the alpha, beta and gamma attributes set to null.</p>
    <div id="log"></div>
    <script>
      var t = async_test("deviceorientationabsolute event fires");
      var run = false;
      window.addEventListener("deviceorientationabsolute", function(e) {
        if (!run) {
          run = true;
          t.step(function () {
            assert_equals(e.type, "deviceorientationabsolute", "type is set to \"deviceorientationabsolute\"");
            assert_equals(e.alpha, null, "alpha is set to null");
            assert_equals(e.beta, null, "beta is set to null");
            assert_equals(e.gamma, null, "gamma is set to null");
            assert_true(e.absolute, "absolute is set to true");
          });
          t.done();
        }
      }, false);
    </script>
  </body>
</html>
